+++ src/smartpeak/include/SmartPeak/core/ApplicationProcessor.h
@@ -71,142 +71,4 @@ namespace SmartPeak
       ISampleGroupProcessorObserver* sample_group_processor_observer = nullptr);
   }
 
-  struct CreateCommand : ApplicationProcessor {
-    CreateCommand(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-    bool process() override;
-    std::string name_;
-    ApplicationHandler::Command cmd_; 
-    virtual std::string getName() const override { return "CreateCommand"; };
-  };
-
-  struct BuildCommandsFromNames : ApplicationProcessor {
-    BuildCommandsFromNames(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-    bool process() override;
-    std::vector<std::string> names_;
-    std::vector<ApplicationHandler::Command> commands_;
-    virtual std::string getName() const override { return "BuildCommandsFromNames"; };
-  };
-
-  struct LoadSession : ApplicationProcessor, IFilePickerHandler
-  {
-    /* IFilePickerHandler */
-    bool onFilePicked(const std::filesystem::path& filename, ApplicationHandler* application_handler) override;
-
-    std::optional<Filenames>    filenames_;           /// Pathnames to load - if not set, read it from session
-    std::optional<Filenames>    filenames_override_;  /// Pathnames override
-    std::optional<ParameterSet> parameters_override_; /// Parameters override
-    std::string      delimiter = ",";                 /// String delimiter of the imported file
-    bool             checkConsistency = true;         /// Check consistency of data contained in files
-
-    LoadSession() = default;
-    explicit LoadSession(
-      ApplicationHandler& application_handler, 
-      WorkflowManager& workflow_manager,
-      IApplicationProcessorObserver* application_processor_observer = nullptr,
-      ISequenceProcessorObserver* sequence_processor_observer = nullptr,
-      ISequenceSegmentProcessorObserver* sequence_segment_processor_observer = nullptr,
-      ISampleGroupProcessorObserver* sample_group_processor_observer = nullptr)
-      : ApplicationProcessor(application_handler),
-        workflow_manager_(workflow_manager),
-        application_processor_observer_(application_processor_observer),
-        sequence_processor_observer_(sequence_processor_observer),
-        sequence_segment_processor_observer_(sequence_segment_processor_observer),
-        sample_group_processor_observer_(sample_group_processor_observer)
-    {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "LOAD_SESSION"; }
-    virtual std::string getDescription() const override { return "Load an existing session"; }
-
-  protected:
-    bool overrideFilenames();
-    bool overrideParameters();
-    bool readFilenames();
-    bool readInputFiles();
-    bool readLoadingWorkflow();
-    bool runLoadingWorkflow();
-    WorkflowManager& workflow_manager_;
-    IApplicationProcessorObserver* application_processor_observer_;
-    ISequenceProcessorObserver* sequence_processor_observer_;
-    ISequenceSegmentProcessorObserver* sequence_segment_processor_observer_;
-    ISampleGroupProcessorObserver* sample_group_processor_observer_;
-  };
-
-  struct LoadPropertiesHandlers : ApplicationProcessor
-  {
-    LoadPropertiesHandlers() = default;
-    explicit LoadPropertiesHandlers(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "LOAD_PROPERTIES_HANDLERS"; }
-    virtual std::string getDescription() const override { return "Load a list of PropertiesHandlers."; }
-
-    std::vector<IPropertiesHandler*> properties_handlers;
-  };
-
-  struct SavePropertiesHandlers : ApplicationProcessor
-  {
-    SavePropertiesHandlers() = default;
-    explicit SavePropertiesHandlers(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "SAVE_PROPERTIES_HANDLERS"; }
-    virtual std::string getDescription() const override { return "Save a list of PropertiesHandlers."; }
-
-    std::vector<IPropertiesHandler*> properties_handlers;
-  };
-
-  struct SaveSession : ApplicationProcessor, IFilePickerHandler
-  {
-    SaveSession() = default;
-    explicit SaveSession(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IFilePickerHandler */
-    bool onFilePicked(const std::filesystem::path& filename, ApplicationHandler* application_handler) override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "SAVE_SESSION"; }
-    virtual std::string getDescription() const override { return "Save the session"; }
-  };
-
-  struct LoadFilenames : ApplicationProcessor
-  {
-    LoadFilenames() = default;
-    explicit LoadFilenames(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "LOAD_FILENAMES"; }
-    virtual std::string getDescription() const override { return "Load Filenames from the DB"; }
-
-    static std::optional<Filenames> loadFilenamesFromDB(const std::filesystem::path& path_db);
-  };
-
-  struct StoreFilenames : ApplicationProcessor
-  {
-    StoreFilenames() = default;
-    explicit StoreFilenames(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
-
-    /* ApplicationProcessor */
-    bool process() override;
-
-    /* IProcessorDescription */
-    virtual std::string getName() const override { return "STORE_FILENAMES"; }
-    virtual std::string getDescription() const override { return "Store Filenames to the DB"; }
-  };
-
 }
