+++ src/smartpeak/include/SmartPeak/core/ApplicationProcessors/BuildCommandsFromNames.h
@@ -23,40 +23,17 @@
 
 #pragma once
 
-#include <SmartPeak/ui/Widget.h>
-#include <SmartPeak/ui/WorkflowStepWidget.h>
-#include <SmartPeak/core/ApplicationHandler.h>
-#include <SmartPeak/core/WorkflowManager.h>
 #include <SmartPeak/core/ApplicationProcessor.h>
-#include <SmartPeak/core/ApplicationProcessors/BuildCommandsFromNames.h>
-#include <string>
-#include <vector>
 
 namespace SmartPeak
 {
-  class WorkflowWidget : public Widget
-  {
-  public:
 
-    WorkflowWidget(const std::string title, ApplicationHandler& application_handler, WorkflowManager& workflow_manager)
-      : Widget(title),
-      application_handler_(application_handler),
-      workflow_step_widget_(application_handler),
-      workflow_manager_(workflow_manager),
-      buildCommandsFromNames_(application_handler)
-    {
-    };
-
-    void draw() override;
-
-  protected:
-    virtual void updatecommands();
-
-  protected:
-    ApplicationHandler& application_handler_;
-    WorkflowStepWidget workflow_step_widget_;
-    WorkflowManager& workflow_manager_;
-    BuildCommandsFromNames buildCommandsFromNames_;
-    bool error_building_commands_ = false;
+  struct BuildCommandsFromNames : ApplicationProcessor {
+    BuildCommandsFromNames(ApplicationHandler& application_handler) : ApplicationProcessor(application_handler) {}
+    bool process() override;
+    std::vector<std::string> names_;
+    std::vector<ApplicationHandler::Command> commands_;
+    virtual std::string getName() const override { return "BuildCommandsFromNames"; };
   };
+
 }
