+++ src/smartpeak/include/SmartPeak/pipelines/GCMS_SIM_Unknown_example.h
@@ -37,6 +37,7 @@
 #include <SmartPeak/core/RawDataProcessors/QuantifyFeatures.h>
 #include <SmartPeak/core/RawDataProcessors/CheckFeatures.h>
 #include <SmartPeak/core/RawDataProcessors/StoreFeatures.h>
+#include <SmartPeak/core/ApplicationProcessors/LoadSession.h>
 
 using namespace SmartPeak;
