+++ src/smartpeak/include/SmartPeak/pipelines/LCMS_MRM_Unknown_example.h
@@ -36,6 +36,7 @@
 #include <SmartPeak/core/RawDataProcessors/QuantifyFeatures.h>
 #include <SmartPeak/core/RawDataProcessors/CheckFeatures.h>
 #include <SmartPeak/core/RawDataProcessors/StoreFeatures.h>
+#include <SmartPeak/core/ApplicationProcessors/LoadSession.h>
 
 using namespace SmartPeak;
