+++ src/smartpeak/source/ui/LoadSessionWizard.cpp
@@ -23,6 +23,8 @@
 
 #include <SmartPeak/ui/LoadSessionWizard.h>
 #include <SmartPeak/core/ApplicationProcessor.h>
+#include <SmartPeak/core/ApplicationProcessors/LoadSession.h>
+#include <SmartPeak/core/ApplicationProcessors/LoadFilenames.h>
 #include <plog/Log.h>
 
 namespace SmartPeak
