+++ src/smartpeak/source/ui/RunWorkflowWidget.cpp
@@ -23,6 +23,7 @@
 
 #include <SmartPeak/ui/RunWorkflowWidget.h>
 #include <SmartPeak/core/ApplicationProcessor.h>
+#include <SmartPeak/core/ApplicationProcessors/BuildCommandsFromNames.h>
 #include <imgui.h>
 #include <misc/cpp/imgui_stdlib.h>
 #include <plog/Log.h>
